feat, fix: 폴더 구조 변경 및 테스트를 위한 생성자에서 받아오는 데이터 변경, ui 구현 완료
