+++ lib/core/routing/router.dart
@@ -56,8 +56,9 @@ final routerProvider = Provider<GoRouter>((ref) {
       GoRoute(
         path: Routes.createProblem,
         builder:
-            (context, state) =>
-                const CreateProblemScreenRoot('clean text test'),
+            (context, state) => const CreateProblemScreenRoot('''
+An important advantage of disclosure, as opposed to more aggressive forms of regulation, is its flexibility and respect for the operation of free markets. Regulatory mandates are blunt swords; they tend to neglect diversity and may have serious unintended adverse effects. For example, energy efficiency requirements for appliances may produce goods that work less well or that have characteristics that consumers do not want. Information provision, by contrast, respects freedom of choice. If automobile manufacturers are required to measure and publicize the safety characteristics of cars, potential car purchasers can trade safety concerns against other attributes, such as price and styling. If restaurant customers are informed of the calories in their meals, those who want to lose weight can make use of the information, leaving those who are unconcerned about calories unaffected. Disclosure does not interfere with, and should even promote, the autonomy (and quality) of individual decision-making.
+'''),
       ),
     ],
   );
